pyproject.toml
[tool.poetry]
name = "pubmed-nonacademic-fetcher"
version = "0.1.0"
description = "Fetch PubMed papers with non-academic authors"
authors = ["Anamika Singh <anamikasingh16052000@gmail.com>"]
readme = "README.md"
packages = [{ include = "get_papers" }]

[tool.poetry.dependencies]
python = ">=3.10"
biopython = "^1.83"
pandas = "^2.2.2"
tabulate = "^0.9.0"

[tool.poetry.scripts]
get-papers-list = "get_papers.cli:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
